LPS-75305 Move to new package
