+++ modules/apps/foundation/petra/petra-concurrent/src/main/java/com/liferay/petra/concurrent/DefaultNoticeableFuture.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.concurrent;
+package com.liferay.petra.concurrent;
 
 import java.util.Set;
 import java.util.concurrent.Callable;
